move logic to blob helper
